feat(console): add `CheckboxGroup`
