+++ packages/console/src/components/Checkbox/CheckboxGroup/index.module.scss
@@ -1,0 +1,7 @@
+@use '@/scss/underscore' as _;
+
+.group {
+  > div:not(:last-child) {
+    margin-bottom: _.unit(3);
+  }
+}
